feat(seatmanager): wait in the queue if other are waiting even seats available for queue first strategy
Add `queue:waiting:count` redis value to record how many parties are
waiting in the queue.

QueueFirstStrategy will assign new arrival party to waiting even there
are enough available seats for that party when there are other parties
are waiting in the queue, that's fair, so renaming the strategy to
FairOrderStrategy.